+++ app/_apis/project.ts
@@ -1,5 +1,6 @@
 import { getHeaders } from "../_constants/HeaderToken";
 import httpClient from "./httpClient";
+import { EditProjectResponse } from "./schema/editProjectResponse";
 import { ProjectResponse, TeamMemberResponse, TotalRatingResponse } from "./schema/projectResponse";
 
 export const projectApi = {
@@ -7,6 +8,11 @@ export const projectApi = {
     const HEADER = getHeaders();
     return await httpClient().get<ProjectResponse>(`/projects/${projectId}`, {}, HEADER.headers, ["projectDetail"]);
   },
+  deleteProject: async (projectId: number) => {
+    const HEADER = getHeaders();
+    return await httpClient().delete(`/projects/${projectId}`, HEADER.headers);
+  },
+
   getTeamMember: async (projectId: number) => {
     return await httpClient().get<TeamMemberResponse>(`/projects/${projectId}/teammates`, {}, {}, [
       "projectTeamMember",
@@ -15,12 +21,26 @@ export const projectApi = {
   getTotalRating: async (projectId: number) => {
     return await httpClient().get<TotalRatingResponse>(`/projects/${projectId}/average-rating`);
   },
-  deleteProject: async (projectId: number) => {
-    const HEADER = getHeaders();
-    return await httpClient().delete(`/projects/${projectId}`, HEADER.headers);
-  },
   postProjectView: async (projectId: number) => {
     const HEADER = getHeaders();
     return await httpClient().postData(`/${projectId}/views`, {}, HEADER.headers);
+  },
+};
+
+export const addProjectApi = {
+  postProject: async (projectData: FormData) => {
+    const HEADER = getHeaders();
+    return await httpClient().postFormData("/projects", projectData, HEADER.headers);
+  },
+};
+
+export const editProjectApi = {
+  putProject: async (projectId: number, projectData: FormData) => {
+    const HEADER = getHeaders();
+    return await httpClient().putFormData(`/projects/${projectId}`, projectData, HEADER.headers);
+  },
+  getProject: async (projectId: number) => {
+    const HEADER = getHeaders();
+    return await httpClient().get<EditProjectResponse>(`/projects/${projectId}/edits`, {}, HEADER.headers);
   },
 };
