+++ app/_queryFactory/editProjectQuery.ts
@@ -1,5 +1,5 @@
 import { createQueryKeys } from "@lukemorales/query-key-factory";
-import { editProjectApi } from "../_apis/editProjectApi";
+import { editProjectApi } from "../_apis/projectApi";
 
 export const editProjectQueryKeys = createQueryKeys("editProject", {
   detail: (projectId: number) => ({
